updated types and consolidated into chart
